commented and cleaned

+++ portell.py
@@ -27,29 +27,44 @@ logger.addHandler(hdlr)
 logger.setLevel(logging.WARNING)
 
 # Set of (protocol, local port) tuples.
-watched = {('tcp4', 5900)}
-sleep_time = 5  # sleep time between checks in seconds
+watched = {('tcp4', 22), ('tcp4', 80), ('tcp4', 5900)}
+
+# time interval between checks for new connections
+sleep_time = 5
+
+# Path variables for the notification application of your choice
 applications = ''.join([expanduser('~'), '/Applications'])
 notifier = '/terminal-notifier.app/Contents/MacOS/terminal-notifier'
 
 while True:
     # Check if any of the watched services is running.
+    # TODO: I'd like this to be streamed instead of poll based, however that
+    # requires moving away from netstat, which is a booger on OSX. Can use
+    # /proc on Linux.
     netstat = check_output(
         [find_executable('netstat'), '-p', 'tcp', '-n'],
         universal_newlines=True).split('\n')
 
     for line in netstat[2:-1]:
-        items = line.split()
-        proto = items[0]
-        port = int(items[3].split('.')[-1])
-        if (proto, port) in watched:
-            logger.warning(line)
-            print line
-            # found = "Found {} connection from {} to port {}"
-            #   .format(proto, items[4], port)
-            found = "You've got mail!"
-            call(
-                [''.join([applications, notifier]),
-                 '-execute', ''.join(["kill -9 ", str(getpid())]),
-                 '-message', found])
+        if line:
+            items = line.split()
+            proto = items[0]
+            port = int(items[3].split('.')[-1])
+            if (proto, port) in watched:
+                # log found connection to log file
+                logger.warning(line)
+
+                # log found connection to stdout
+                print line
+
+                # format alert notification text
+                alertText = "Found {} connection from {} to port {}"
+                alertFormatted = alertText.format(proto, items[4], port)
+
+                # send formatted alert notification text to terminal-notifer
+                # with click to quit options set.
+                call(
+                    [''.join([applications, notifier]),
+                     '-execute', ''.join(["kill -9 ", str(getpid())]),
+                     '-message', alertFormatted])
     sleep(sleep_time)
